Moved the wcs utility functions from gammapy.background.models to gammapy.utils.wcs.
