+++ docs/utils/index.rst
@@ -56,3 +56,6 @@ Reference/API
 
 .. automodapi:: gammapy.utils.testing
     :no-inheritance-diagram:
+
+.. automodapi:: gammapy.utils.wcs
+    :no-inheritance-diagram:
