+++ gammapy/background/tests/test_models.py
@@ -10,7 +10,6 @@ from astropy.units import Quantity
 from astropy.coordinates import Angle
 from astropy.modeling.models import Gaussian1D
 from ...background import GaussianBand2D, CubeBackgroundModel
-from ...background import make_linear_bin_edges_arrays_from_wcs, make_linear_wcs_from_bin_edges_arrays
 from ... import datasets
 
 
@@ -50,21 +49,6 @@ class TestGaussianBand2D():
         model = self.model.y_model(-30)
         assert isinstance(model, Gaussian1D)
         assert_allclose(model.parameters, [0, -1, 0.4])
-
-
-def test_wcs_object():
-    bins_x = Angle([1., 2., 3., 4.], 'degree')
-    bins_y = Angle([-1.5, 0., 1.5], 'degree')
-    wcs = make_linear_wcs_from_bin_edges_arrays("X", "Y", bins_x, bins_y)
-    nbins_x = len(bins_x) - 1
-    nbins_y = len(bins_y) - 1
-    reco_bins_x, reco_bins_y = make_linear_bin_edges_arrays_from_wcs(wcs,
-                                                                     nbins_x,
-                                                                     nbins_y)
-
-    # test: reconstructed bins should match original bins
-    assert_allclose(reco_bins_x, bins_x)
-    assert_allclose(reco_bins_y, bins_y)
 
 
 class TestCubeBackgroundModel():
