+++ gammapy/utils/__init__.py
@@ -4,6 +4,6 @@ Utility functions and classes that are used in many places
 or didn't fit anywhere else.
 
 You have to import sub-modules like `gammapy.utils.root` or
-`gammapy.utils.random` or `gammapy.utils.
+`gammapy.utils.random` or `gammapy.utils.wcs`.
 Importing `gammapy.utils` directly is not useful.
 """
